Adminitrative Section added
Added a new section (The Administration Section). It holds the name,
title, and email of the Administrative team! This section is found in the
parents section of the page just in case they wish to get in contact with
any of the individuals from the administrative team (lodging complaints,
feedback, etc).